+++ PKHeX.Core/Legality/Encounters/Templates/Gen4/IGroundTypeTile.cs
@@ -18,5 +18,5 @@ public interface IGroundTypeTile
     /// Gets if the resulting <see cref="PKM"/> will still have a value depending on the current <see cref="format"/>.
     /// </summary>
     /// <remarks>Generation 7 no longer stores this value.</remarks>
-    bool HasGroundTile(int format) => format is (4 or 5 or 6);
+    bool HasGroundTile(byte format) => format is (4 or 5 or 6);
 }
