+++ PKHeX.Core/Legality/LearnSource/Group/LearnGroup6.cs
@@ -8,7 +8,7 @@ namespace PKHeX.Core;
 public sealed class LearnGroup6 : ILearnGroup
 {
     public static readonly LearnGroup6 Instance = new();
-    private const int Generation = 6;
+    private const byte Generation = 6;
     public ushort MaxMoveID => Legal.MaxMoveID_6_AO;
 
     public ILearnGroup? GetPrevious(PKM pk, EvolutionHistory history, IEncounterTemplate enc, LearnOption option) => enc.Generation is Generation ? null : LearnGroup5.Instance;
