+++ PKHeX.Core/Legality/LearnSource/Sources/LearnSource4DP.cs
@@ -15,7 +15,7 @@ public sealed class LearnSource4DP : LearnSource4, ILearnSource<PersonalInfo4>, 
     private static readonly Learnset[] Learnsets = LearnsetReader.GetArray(BinLinkerAccessor.Get(Util.GetBinaryResource("lvlmove_dp.pkl"), "dp"u8));
     private const int MaxSpecies = Legal.MaxSpeciesID_4;
     private const LearnEnvironment Game = DP;
-    private const int Generation = 4;
+    private const byte Generation = 4;
     private const int CountTM = 92;
 
     public Learnset GetLearnset(ushort species, byte form) => Learnsets[Personal.GetFormIndex(species, form)];
